changchang

+++ serial_program.cpp
@@ -8,11 +8,11 @@ using namespace std;
 
 int IT_NUM = 10;
 
-int func(int a, int b, int c, int d) {
+long long func(long long a, long long b, long long c, long long d) {
     return (a + b + c + d) % 100;
 }
 
-void run_serial(vector<vector<int>>& A0, int n, int (*f)(int, int, int, int), int if_print) {
+void run_serial(vector<vector<int>>& A0, int n, long long (*f)(long long, long long, long long, long long), int if_print) {
     cout << "n is: " << n << endl;
 
     double start = MPI_Wtime();
